Move review app functions to their own files  🐿 v2.10.3
