+++ lib/github-api.js
@@ -1,0 +1,25 @@
+const getGithubArchiveUrl = ({ repoName, branch }) => `https://api.github.com/repos/Financial-Times/${repoName}/tarball/${branch}`;
+
+const getGithubArchiveRedirectUrl = ({ repoName, branch, githubToken }) => {
+	const url = getGithubArchiveUrl({ repoName, branch });
+
+	return fetch(url, {
+		headers: {
+			Authorization: `token ${githubToken}`
+		},
+		redirect: 'manual' // Don't follow redirect, just want the URL
+	}).then(res => {
+		if (res.status !== 302) {
+			throw new Error(`Unexpected response for ${url} (${status})`);
+		}
+
+		const { headers: { _headers: { location } } } = res;
+		const [ redirectUrl ] = location || [];
+
+		return redirectUrl;
+	});
+};
+
+module.exports = {
+	getGithubArchiveRedirectUrl
+};
