+++ tasks/review-app.js
@@ -1,204 +1,17 @@
-const pRetry = require('p-retry');
-
-const herokuAuthToken = require('../lib/heroku-auth-token');
 const { info: pipelineInfo } = require('../lib/pipelines');
-const REVIEW_APPS_URL = 'https://api.heroku.com/review-apps';
+const {
+	createReviewApp,
+	findCreatedReviewApp,
+	waitTillReviewAppCreated,
+	waitForReviewAppBuild,
+	getAppName
+} = require('../lib/review-apps');
 
-const DEFAULT_HEADERS = {
-	'Accept': 'application/vnd.heroku+json; version=3',
-	'Content-Type': 'application/json'
-};
-
-const NUM_RETRIES = 30;
-const RETRY_EXP_BACK_OFF_FACTOR = 1;
-const RETRY_INTERVAL = 10 * 1000;
-const REVIEW_APP_STATUSES = {
-	pending: 'pending',
-	deleted: 'deleted',
-	creating: 'creating',
-	created: 'created'
-};
-const BUILD_STATUS_SUCCEEDED = 'succeeded';
-
-const getReviewAppUrl = reviewAppId => `https://api.heroku.com/review-apps/${reviewAppId}`;
-const getPipelineReviewAppsUrl = pipelineId => `https://api.heroku.com/pipelines/${pipelineId}/review-apps`;
-const getAppUrl = appId => `https://api.heroku.com/apps/${appId}`;
-const getGithubArchiveUrl = ({ repoName, branch }) => `https://api.github.com/repos/Financial-Times/${repoName}/tarball/${branch}`;
-const getBuildsUrl = appId => `https://api.heroku.com/apps/${appId}/builds`;
-
-function herokuHeaders ({ useReviewAppApi } = {}) {
-	const defaultHeaders = useReviewAppApi
-		? Object.assign({}, DEFAULT_HEADERS, {
-			Accept: 'application/vnd.heroku+json; version=3.review-apps',
-		})
-		: DEFAULT_HEADERS;
-	return herokuAuthToken()
-		.then(key => {
-			return {
-				...defaultHeaders,
-				Authorization: `Bearer ${key}`
-			};
-		});
-}
-
-const throwIfNotOk = async res => {
-	const { ok, status, url } = res;
-	if (!ok) {
-		const errorBody = await res.json();
-
-		console.error('Fetch error:', status, url, errorBody); // eslint-disable-line no-console
-		throw errorBody;
-	}
-	return res;
-};
-
-const getGithubArchiveRedirectUrl = ({ repoName, branch, githubToken }) => {
-	const url = getGithubArchiveUrl({ repoName, branch });
-
-	return fetch(url, {
-		headers: {
-			Authorization: `token ${githubToken}`
-		},
-		redirect: 'manual' // Don't follow redirect, just want the URL
-	}).then(res => {
-		if (res.status !== 302) {
-			throw new Error(`Unexpected response for ${url} (${status})`);
-		}
-
-		const { headers: { _headers: { location } } } = res;
-		const [ redirectUrl ] = location || [];
-
-		return redirectUrl;
-	});
-};
-
-const waitTillReviewAppCreated = (data) => {
-	const { id } = data;
-	const checkForCreatedStatus = async () => {
-		const headers = await herokuHeaders({ useReviewAppApi: true });
-		const result = await fetch(getReviewAppUrl(id), {
-			headers
-		})
-			.then(throwIfNotOk)
-			.then(res => res.json())
-			.then(data => {
-				const { status, message, app } = data;
-
-				if (status === REVIEW_APP_STATUSES.deleted) {
-					throw new pRetry.AbortError(`Review app was deleted: ${message}`);
-				}
-
-				if (status !== REVIEW_APP_STATUSES.created) {
-					const appIdOutput = (status === REVIEW_APP_STATUSES.creating)
-						? `, appId: ${app.id}`
-						: '';
-					throw new Error(`Review app not created yet. Current status: ${status}${appIdOutput}`);
-				};
-
-				return app.id;
-			});
-		return result;
-	};
-
-	return pRetry(checkForCreatedStatus, {
-		factor: RETRY_EXP_BACK_OFF_FACTOR,
-		retries: NUM_RETRIES,
-		minTimeout: RETRY_INTERVAL,
-		onFailedAttempt: (err) => {
-			const { attemptNumber, message } = err;
-			console.error(`${attemptNumber}/${NUM_RETRIES}: ${message}`); // eslint-disable-line no-console
-		}
-	});
-};
-
-const getAppName = async (appId) => {
-	const headers = await herokuHeaders();
-	return fetch(getAppUrl(appId), {
-		headers
-	})
-		.then(throwIfNotOk)
-		.then(res => res.json())
-		.then((result) => {
-			const { name } = result;
-			return name;
-		});
-};
-
-const findCreatedReviewApp = async ({ pipelineId, branch }) => {
-	const headers = await herokuHeaders({ useReviewAppApi: true });
-	return fetch(getPipelineReviewAppsUrl(pipelineId), {
-		headers
-	})
-		.then(throwIfNotOk)
-		.then(res => res.json())
-		.then((reviewApps = []) =>
-			reviewApps.find(({ branch: reviewAppBranch }) => reviewAppBranch === branch));
-};
-
-const getBuilds = async (appId) => {
-	const headers = await herokuHeaders();
-	return fetch(getBuildsUrl(appId), {
-		headers
-	})
-		.then(throwIfNotOk)
-		.then(res => res.json());
-};
-
-const waitForReviewAppBuild = (commit) => async (appId) => {
-	const checkForBuildAppId = () => getBuilds(appId)
-		.then(builds => {
-			const build = builds.find(({ source_blob: { version } }) => version === commit);
-
-			if (!build) {
-				throw new Error(`No review app build found for app id ${appId}, commit ${commit}`);
-			}
-
-			const { status } = build;
-			if (status !== BUILD_STATUS_SUCCEEDED) {
-				throw new Error(`Review app build for app id ${appId} (commit ${commit}) not done yet: ${status}`);
-			}
-
-			return build;
-		})
-		.then(build => {
-			if (!build) {
-				throw new Error (`Build for commit ${commit} not found`);
-			}
-			return build;
-		})
-		.then(({ app: { id } }) => id);
-
-	return pRetry(checkForBuildAppId, {
-		factor: RETRY_EXP_BACK_OFF_FACTOR,
-		retries: NUM_RETRIES,
-		minTimeout: RETRY_INTERVAL,
-		onFailedAttempt: (err) => {
-			const { attemptNumber, message } = err;
-			console.error(`${attemptNumber}/${NUM_RETRIES}: ${message}`); // eslint-disable-line no-console
-		}
-	});
-};
-
-async function task (app, options) {
+async function task (appName, options) {
 	const { repoName, branch, commit, githubToken } = options;
+	const { id: pipelineId } = await pipelineInfo(appName);
 
-	const { id: pipelineId } = await pipelineInfo(app);
-	const headers = await herokuHeaders({ useReviewAppApi: true });
-	const body = {
-		pipeline: pipelineId,
-		branch,
-		source_blob: {
-			url: await getGithubArchiveRedirectUrl({ repoName, branch, githubToken }),
-			version: commit
-		}
-	};
-	const createReviewApp = () => fetch(REVIEW_APPS_URL, {
-		headers,
-		method: 'post',
-		body: JSON.stringify(body)
-	});
-
-	return createReviewApp()
+	return createReviewApp({ pipelineId, repoName, commit, branch, githubToken })
 		.then(res => {
 			const { status } = res;
 			if (status === 409) {
@@ -219,7 +32,6 @@ async function task (app, options) {
 					.then(getAppName);
 			}
 			return Promise.resolve(res)
-				.then(throwIfNotOk)
 				.then(res => res.json())
 				.then(waitTillReviewAppCreated)
 				.then(getAppName);
@@ -229,19 +41,19 @@ async function task (app, options) {
 		});
 }
 
-const description = 'Create or find an existing heroku review app and print out the app name. [app] is the package.json name (which is also the value of VAULT_SOURCE). On the first build of a branch, Heroku will create a review app with a build. On subsequent builds, Heroku will automatically generate a new build, which this task looks for. See https://devcenter.heroku.com/articles/review-apps-beta for more details of the internals';
+const description = 'Create or find an existing heroku review app and print out the app name. [appName] is the package.json name (which is also the value of VAULT_SOURCE). On the first build of a branch, Heroku will create a review app with a build. On subsequent builds, Heroku will automatically generate a new build, which this task looks for. See https://devcenter.heroku.com/articles/review-apps-beta for more details of the internals';
 
 module.exports = function (program) {
 	program
-		.command('review-app [app]')
+		.command('review-app [appName]')
 		.description(description)
 		.option('-r, --repo-name <name>', 'github repository name')
 		.option('-b, --branch <name>', 'branch of the review app')
 		.option('-c, --commit <commit>', 'commit SHA-1')
 		.option('-g, --github-token <token>', 'github personal token to access source code (generate from https://github.com/settings/tokens)')
-		.action(async function (app, options) {
+		.action(async function (appName, options) {
 			try {
-				await task(app, options);
+				await task(appName, options);
 			} catch (error) {
 				console.error(error); // eslint-disable-line no-console
 				process.exit(1);
